+++ include/sealib/iterator/dfs.h
@@ -13,9 +13,9 @@
 
 namespace Sealib {
 
-const uint DFS_WHITE = 0, DFS_GRAY = 1, DFS_BLACK = 2, DFS_RESERVED = 3;
-const Consumer DFS_NOP_PROCESS = [](uint) {};
-const BiConsumer DFS_NOP_EXPLORE = [](uint, uint) {};
+const uint64_t DFS_WHITE = 0, DFS_GRAY = 1, DFS_BLACK = 2, DFS_RESERVED = 3;
+const Consumer DFS_NOP_PROCESS = [](uint64_t) {};
+const BiConsumer DFS_NOP_EXPLORE = [](uint64_t, uint64_t) {};
 
 /**
  * This class contains depth-first search algorithms.
@@ -108,36 +108,36 @@ class DFS {
      * @param startVertex startVertex The begin of the DFS tree.
      * @author Simon Schniedenharn
      */
-    static void runLinearTimeInplaceDFS(uint32_t *graph, Consumer preProcess,
+    static void runLinearTimeInplaceDFS(uint64_t *graph, Consumer preProcess,
                                         Consumer postProcess,
-                                        uint32_t startVertex);
+                                        uint64_t startVertex);
 
     /**
      * The following helper procedures are only for internal or experimental
      * usage.
      */
 
-    static void visit_standard(uint u0, Graph const *g,
-                               std::vector<uint> *color,
-                               std::vector<std::pair<uint, uint>> *s,
+    static void visit_standard(uint64_t u0, Graph const *g,
+                               std::vector<uint64_t> *color,
+                               std::vector<std::pair<uint64_t, uint64_t>> *s,
                                Consumer preProcess, BiConsumer preExplore,
                                BiConsumer postExplore, Consumer postProcess);
 
-    static void visit_nloglogn(uint u0, Graph const *g, CompactArray *color,
+    static void visit_nloglogn(uint64_t u0, Graph const *g, CompactArray *color,
                                SegmentStack *s,
-                               std::function<void(uint u0)> restoration,
+                               std::function<void(uint64_t u0)> restoration,
                                Consumer preprocess, BiConsumer preexplore,
                                BiConsumer postexplore, Consumer postprocess);
 
-    static void visit_nplusm(uint u0, UndirectedGraph const *g,
+    static void visit_nplusm(uint64_t u0, UndirectedGraph const *g,
                              CompactArray *color, Sequence<uint64_t> *back,
                              Consumer preprocess, BiConsumer preexplore,
                              BiConsumer postexplore, Consumer postprocess);
 
-    static void restore_full(uint u0, Graph const *g, CompactArray *color,
+    static void restore_full(uint64_t u0, Graph const *g, CompactArray *color,
                              BasicSegmentStack *s);
 
-    static void restore_top(uint u0, Graph const *g, CompactArray *color,
+    static void restore_top(uint64_t u0, Graph const *g, CompactArray *color,
                             ExtendedSegmentStack *s);
 };
 }  // namespace Sealib
